Adjust the id to long instead of integer

+++ src/main/java/com/capgemini/molveno/reserveringssysteem/repository/RoomRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Kamer, Integer> {
+public interface RoomRepository extends JpaRepository<Kamer, Long> {
 }
